requirements.txt
flask==3.0.0
werkzeug==3.0.1
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.112.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
reportlab==4.0.8
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dateutil==2.8.2
pydub==0.25.1
pandas==2.1.4
numpy==1.25.2
